图片，添加User-Agent, https

+++ app/src/main/java/com/yxws/msettopboxs/net/requestbean/WatchHistoryBean.java
@@ -4,10 +4,10 @@ public class WatchHistoryBean {
     private String contentName;//视频名称
     private String contentId;//cp_album_id这个ID
     private int extraContentId;//cp_tv_id这个ID
-    private int contentTotalTime;//节目总时长：秒（s）
-    private int startWatchTime;//用户开始看的时间
-    private int endWatchTime;//用户结束看时间
-    private int playTime;//用户看了多少时长
+    private long contentTotalTime;//节目总时长：秒（s）
+    private long startWatchTime;//用户开始看的时间
+    private long endWatchTime;//用户结束看时间
+    private long playTime;//用户看了多少时长
     private String logType;//同步数据时，用户当前播放状态
     private String account;//用户登录账号,机顶盒登陆账号
     private String imageUrl;//海报URL
@@ -36,35 +36,35 @@ public class WatchHistoryBean {
         this.extraContentId = extraContentId;
     }
 
-    public int getContentTotalTime() {
+    public long getContentTotalTime() {
         return contentTotalTime;
     }
 
-    public void setContentTotalTime(int contentTotalTime) {
+    public void setContentTotalTime(long contentTotalTime) {
         this.contentTotalTime = contentTotalTime;
     }
 
-    public int getStartWatchTime() {
+    public long getStartWatchTime() {
         return startWatchTime;
     }
 
-    public void setStartWatchTime(int startWatchTime) {
+    public void setStartWatchTime(long startWatchTime) {
         this.startWatchTime = startWatchTime;
     }
 
-    public int getEndWatchTime() {
+    public long getEndWatchTime() {
         return endWatchTime;
     }
 
-    public void setEndWatchTime(int endWatchTime) {
+    public void setEndWatchTime(long endWatchTime) {
         this.endWatchTime = endWatchTime;
     }
 
-    public int getPlayTime() {
+    public long getPlayTime() {
         return playTime;
     }
 
-    public void setPlayTime(int playTime) {
+    public void setPlayTime(long playTime) {
         this.playTime = playTime;
     }
